python-ai-service/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0

# Database & Queue
asyncpg==0.29.0
sqlalchemy==2.0.25
alembic==1.13.1

# Document Processing
PyMuPDF==1.23.21
python-docx==1.1.0
beautifulsoup4==4.12.3
chardet==5.2.0

# NLP & Text Processing
spacy==3.7.2
nltk==3.8.1
textstat==0.7.3
langdetect==1.0.9

# AI/ML
openai==1.10.0
anthropic==0.19.1
tiktoken==0.5.2
sentence-transformers==2.3.1
torch==2.1.2
numpy==1.26.3
llama-cpp-python==0.2.55
pydantic-settings==2.1.0

# Async & Performance
aiohttp==3.9.3
aiofiles==23.2.1
httpx==0.26.0
redis==5.0.1

# Monitoring & Logging
structlog==24.1.0
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.40.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0

# Development
black==24.1.1
ruff==0.1.14
mypy==1.8.0
pre-commit==3.6.0